Fix Piglin capture
