+++ src/main/java/io/github/thebusybiscuit/mobcapturer/mobs/PiglinAdapter.java
@@ -1,29 +1,16 @@
 package io.github.thebusybiscuit.mobcapturer.mobs;
 
-import java.util.List;
-
 import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 import com.google.gson.JsonObject;
 
-import org.bukkit.ChatColor;
 import org.bukkit.entity.Piglin;
 
-public class PiglinAdapter extends AbstractHumanoidAdapter<Piglin> {
+public class PiglinAdapter extends AbstractPiglinAdapter<Piglin> {
 
     public PiglinAdapter() {
         super(Piglin.class);
-    }
-
-    @Nonnull
-    @Override
-    public List<String> getLore(@Nonnull JsonObject json) {
-        List<String> lore = super.getLore(json);
-
-        lore.add(ChatColor.GRAY + "Baby: " + ChatColor.WHITE + json.get("baby").getAsBoolean());
-
-        return lore;
     }
 
     @ParametersAreNonnullByDefault
@@ -31,10 +18,7 @@ public class PiglinAdapter extends AbstractHumanoidAdapter<Piglin> {
     public void apply(Piglin entity, JsonObject json) {
         super.apply(entity, json);
 
-        entity.setAge(json.get("age").getAsInt());
         entity.setIsAbleToHunt(json.get("ableToHunt").getAsBoolean());
-        entity.setConversionTime(json.get("conversionTime").getAsInt());
-        entity.setImmuneToZombification(json.get("immuneToZombification").getAsBoolean());
     }
 
     @Nonnull
@@ -42,11 +26,7 @@ public class PiglinAdapter extends AbstractHumanoidAdapter<Piglin> {
     public JsonObject saveData(@Nonnull Piglin entity) {
         JsonObject json = super.saveData(entity);
 
-        json.addProperty("age", entity.getAge());
-        json.addProperty("baby", !entity.isAdult());
         json.addProperty("ableToHunt", entity.isAbleToHunt());
-        json.addProperty("conversionTime", entity.getConversionTime());
-        json.addProperty("immuneToZombification", entity.isImmuneToZombification());
 
         return json;
     }
